refactor to util component IntlMessageExtractor
